+++ crates/adapters/okx/bin/http_public.rs
@@ -24,11 +24,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Request instruments
     let inst_type = OKXInstrumentType::Swap;
-    let instruments = client.request_instruments(inst_type, None).await?;
+    let (instruments, _inst_id_codes) = client.request_instruments(inst_type, None).await?;
     client.cache_instruments(instruments);
 
     let inst_type = OKXInstrumentType::Spot;
-    let instruments = client.request_instruments(inst_type, None).await?;
+    let (instruments, _inst_id_codes) = client.request_instruments(inst_type, None).await?;
     client.cache_instruments(instruments);
 
     // Request mark price